Move all classes into bob package
