+++ docs/README.md
@@ -1,4 +1,4 @@
-# Bob User Guide
+# bob.Bob User Guide
 
 // Update the title above to match the actual product name
