+++ src/main/java/bob/Bob.java
@@ -1,3 +1,5 @@
+package bob;
+
 public class Bob {
     private Storage storage;
     private TaskList tasks;
