+++ src/main/java/bob/Command.java
@@ -1,3 +1,5 @@
+package bob;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
