+++ src/main/java/bob/Deadline.java
@@ -1,3 +1,5 @@
+package bob;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
