+++ src/main/java/bob/Event.java
@@ -1,3 +1,5 @@
+package bob;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
