+++ src/main/java/bob/ExitCommand.java
@@ -1,3 +1,5 @@
+package bob;
+
 class ExitCommand extends Command {
     protected static String[] params = new String[] { "bye" };
     protected static int paramCount = 0;
