+++ src/main/java/bob/MissingParamException.java
@@ -1,3 +1,5 @@
+package bob;
+
 public class MissingParamException  extends IllegalInputException {
     public MissingParamException (String offender) {
         super("Your instructions were missing some paramters:",
