+++ src/main/java/bob/Storage.java
@@ -1,3 +1,5 @@
+package bob;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
