+++ src/main/java/bob/Task.java
@@ -1,3 +1,5 @@
+package bob;
+
 abstract public class Task {
     private String name;
     private boolean completed = false;
