+++ src/main/java/bob/TaskIndexException.java
@@ -1,3 +1,5 @@
+package bob;
+
 public class TaskIndexException extends IllegalInputException {
     public TaskIndexException(String offender) {
         super("The index you have selected is not allowed:",
