+++ src/main/java/bob/TodoCommand.java
@@ -1,3 +1,5 @@
+package bob;
+
 public class TodoCommand  extends Command {
     protected static String[] params = new String[] { "todo" };
     protected static int paramCount = 1;
