+++ src/main/java/bob/Ui.java
@@ -1,7 +1,9 @@
+package bob;
+
 import java.util.Scanner;
 
 public class Ui {
-    private static final String[] welcome = { "Hello! I'm Bob", "What can I do for you?" };
+    private static final String[] welcome = { "Hello! I'm bob.Bob", "What can I do for you?" };
     private static final String[] farewell = { " Bye. Hope to see you again soon!" };
 
     public void showWelcome() {
