+++ src/main/java/bob/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package bob;
+
 public class UnknownCommandException extends IllegalInputException {
     public UnknownCommandException(String offender) {
         super("I'm unable to understand this instruction:",
