+++ src/main/java/bob/UnmarkCommand.java
@@ -1,18 +1,20 @@
-public class MarkCommand  extends Command {
+package bob;
+
+public class UnmarkCommand  extends Command {
     protected static String[] params = new String[] { "mark" };
     protected static int paramCount = 1;
     protected static String identifier = "mark";
     private final int idx;
 
-    public MarkCommand(int idx) {
+    public UnmarkCommand(int idx) {
         this.idx = idx;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.mark(idx);
+        tasks.unmark(idx);
         Printer.prettyPrint(new String[] {
-                "Nice! I've marked this task as done:",
+                "OK, I've marked this task as not done yet:",
                 tasks.describeTask(idx) });
     }
 }
