+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-     Hello! I'm Bob
+     Hello! I'm bob.Bob
      What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
